+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix API marshalling errors with system checks and app component types.
 * You can now refer to system checks by type in commands, e.g. `system check delete my.cool.system disk`
 * You can now enable or disable DKIM on mail domains with `lvl mail domain dkim`.
+* `lvl login` now supports `--trace` and passes the correct `User-Agent` to the API.
 
 ## 1.6.0